requirements.txt
beautifulsoup4==4.4.1
cookies==2.2.1
Flask==0.10.1
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mxshipalerts==0.0.1
pdfkit==0.5.0
py==1.4.31
pytest==2.8.7
PyYAML==3.11
requests==2.9.1
responses==0.5.1
six==1.10.0
vcrpy==1.7.4
Werkzeug==0.11.3
wrapt==1.10.6
